Add basic styled html display and basic js objects
The display contains the game board and an interface to display
scores, game messages, and buttons to reset scores and start a new
game. The js file contains the beginnings of objects created from
factories and modules to represent players, the game board, the game
state and the interface state.